feat,refactory (model, repository,service):Add the class service and change details in class_repo as in the model

+++ internal/model/classes.go
@@ -1,7 +1,7 @@
 package model
 
 type Classes struct {
-	ID       int16  `json:"id_class"`
+	ID       int    `json:"id_class"`
 	Name     string `json:"class_name"`
 	Profesor int16  `json:"class_profesor"`
 	Curso    string `json:"class_curso"`
